# backend/src/main/resources/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Tracking what you have watched and where you left off, for films and TV series.\n\nEvery endpoint under /api/v1 requires an OIDC bearer token from a trusted issuer (Sign in with Apple or Google). The account is created on the first valid token, so there is no registration call. Endpoints act on the user the token identifies -- there is no way to name a different one.\n\nCatalog metadata comes from TMDB. This product uses the TMDB API but is not endorsed or certified by TMDB.",
    "title" : "Watch Guru API",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/me/watchlist" : {
      "get" : {
        "operationId" : "listWatchlist",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "WATCHLIST", "WATCHING", "COMPLETED", "ON_HOLD", "DROPPED" ]
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/WatchlistItemResponse"
                  }
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "watchlist-controller" ]
      },
      "post" : {
        "operationId" : "addToWatchlist",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AddToWatchlist"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WatchlistItemResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "watchlist-controller" ]
      }
    },
    "/api/v1/me/watchlist/titles/{titleId}/progress" : {
      "get" : {
        "operationId" : "getTitleProgress",
        "parameters" : [ {
          "in" : "path",
          "name" : "titleId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TitleProgress"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "watchlist-controller" ]
      }
    },
    "/api/v1/me/watchlist/{itemId}" : {
      "delete" : {
        "operationId" : "removeFromWatchlist",
        "parameters" : [ {
          "in" : "path",
          "name" : "itemId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "watchlist-controller" ]
      },
      "patch" : {
        "operationId" : "updateWatchlistItem",
        "parameters" : [ {
          "in" : "path",
          "name" : "itemId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateWatchlistItem"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WatchlistItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "watchlist-controller" ]
      }
    },
...
    "/api/v1/titles/import" : {
      "post" : {
        "operationId" : "importTitle",
        "parameters" : [ {
          "in" : "query",
          "name" : "titleType",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "MOVIE", "TV_SERIES" ]
          }
        }, {
          "in" : "query",
          "name" : "providerId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "language",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TitleResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "title-controller" ]
      }
    },
    "/api/v1/titles/search" : {
      "get" : {
        "operationId" : "searchTitles",
        "parameters" : [ {
          "in" : "query",
          "name" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "minLength" : 1
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "in" : "query",
          "name" : "language",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SearchResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "title-controller" ]
      }
    },
    "/api/v1/titles/{titleId}" : {
      "get" : {
        "operationId" : "getTitle",
        "parameters" : [ {
          "in" : "path",
          "name" : "titleId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "in" : "query",
          "name" : "region",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TitleResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "title-controller" ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "AddToWatchlist" : {
        "type" : "object",
        "properties" : {
          "providerId" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          },
          "status" : {
            "type" : "string",
            "enum" : [ "WATCHLIST", "WATCHING", "COMPLETED", "ON_HOLD", "DROPPED" ]
          },
          "titleType" : {
            "type" : "string",
            "enum" : [ "MOVIE", "TV_SERIES" ]
          }
        },
        "required" : [ "providerId", "titleType" ]
      },
      "AvailabilityResponse" : {
        "type" : "object",
        "properties" : {
          "fetchedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "link" : {
            "type" : "string"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "offerType" : {
            "type" : "string",
            "enum" : [ "FLATRATE", "RENT", "BUY", "ADS", "FREE" ]
          },
          "serviceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "serviceName" : {
            "type" : "string"
          }
        },
        "required" : [ "fetchedAt", "offerType", "serviceId", "serviceName" ]
      },
...
      "GenreResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
...
      "SearchHit" : {
        "type" : "object",
        "properties" : {
          "originalTitle" : {
            "type" : "string"
          },
          "overview" : {
            "type" : "string"
          },
          "posterUrl" : {
            "type" : "string"
          },
          "providerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "providerRating" : {
            "type" : "number"
          },
          "providerVoteCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "releaseDate" : {
            "type" : "string",
            "format" : "date"
          },
          "title" : {
            "type" : "string"
          },
          "titleType" : {
            "type" : "string",
            "enum" : [ "MOVIE", "TV_SERIES" ]
          }
        },
        "required" : [ "providerId", "title", "titleType" ]
      },
      "SearchResponse" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "results" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SearchHit"
            }
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalResults" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "page", "results", "totalPages", "totalResults" ]
      },
...
      "TitleProgress" : {
        "type" : "object",
        "properties" : {
          "airedEpisodes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nextEpisodeCode" : {
            "type" : "string"
          },
          "nextEpisodeId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nextEpisodeName" : {
            "type" : "string"
          },
          "percentComplete" : {
            "type" : "integer",
            "format" : "int32"
          },
          "primaryTitle" : {
            "type" : "string"
          },
          "remainingMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "titleId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "watchedEpisodes" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "airedEpisodes", "percentComplete", "primaryTitle", "remainingMinutes", "titleId", "watchedEpisodes" ]
      },
      "TitleResponse" : {
        "type" : "object",
        "properties" : {
          "availability" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AvailabilityResponse"
            }
          },
          "backdropUrl" : {
            "type" : "string"
          },
          "genres" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GenreResponse"
            }
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "imdbId" : {
            "type" : "string"
          },
          "imdbRating" : {
            "type" : "number"
          },
          "imdbUrl" : {
            "type" : "string"
          },
          "lastAirDate" : {
            "type" : "string",
            "format" : "date"
          },
          "numberOfEpisodes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "numberOfSeasons" : {
            "type" : "integer",
            "format" : "int32"
          },
          "originalTitle" : {
            "type" : "string"
          },
          "overview" : {
            "type" : "string"
          },
          "posterUrl" : {
            "type" : "string"
          },
          "primaryTitle" : {
            "type" : "string"
          },
          "productionStatus" : {
            "type" : "string"
          },
          "providerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "providerRating" : {
            "type" : "number"
          },
          "providerVoteCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "releaseDate" : {
            "type" : "string",
            "format" : "date"
          },
          "runtimeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tagline" : {
            "type" : "string"
          },
          "titleType" : {
            "type" : "string",
            "enum" : [ "MOVIE", "TV_SERIES" ]
          }
        },
        "required" : [ "availability", "genres", "id", "primaryTitle", "providerId", "titleType" ]
      },
...
      "UpdateWatchlistItem" : {
        "type" : "object",
        "properties" : {
          "notes" : {
            "type" : "string"
          },
          "rating" : {
            "type" : "number",
            "maximum" : 10.0,
            "minimum" : 0.0
          },
          "status" : {
            "type" : "string",
            "enum" : [ "WATCHLIST", "WATCHING", "COMPLETED", "ON_HOLD", "DROPPED" ]
          }
        }
      },
...
      "WatchlistItemResponse" : {
        "type" : "object",
        "properties" : {
          "addedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "completedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "favorite" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "notes" : {
            "type" : "string"
          },
          "priority" : {
            "type" : "integer",
            "format" : "int32"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "WATCHLIST", "WATCHING", "COMPLETED", "ON_HOLD", "DROPPED" ]
          },
          "title" : {
            "$ref" : "#/components/schemas/TitleResponse"
          },
          "userRating" : {
            "type" : "number"
          }
        },
        "required" : [ "addedAt", "favorite", "id", "priority", "status", "title" ]
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "OIDC ID token from Sign in with Apple or Google.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}